io: move the S3 request-signing files under io/rest/
Relocation only, no content change. SigV4 signing, the list parser and
the credential/authorizer types are consumed by the REST reactor rather
than being an S3 backend of their own, so they belong beside it:

  io/s3/s3_list_parser          -> io/rest/s3/list_parser
  io/s3/sigv4                   -> io/rest/s3/sigv4
  io/s3/sirius_sigv4_authorizer -> io/rest/s3/sigv4_authorizer
  io/s3/static_credentials      -> io/rest/s3/static_credentials
  io/s3/mock_request_authorizer -> io/rest/mock_authorizer
  io/s3/s3_request_authorizer   -> io/rest/authorizer

sirius_httpfs stays put -- it is a DuckDB filesystem, not part of the
reactor's signing path. Include paths updated mechanically.
