+++ src/include/io/rest/config.hpp
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include "io/s3/s3_list_parser.hpp"
+#include "io/rest/s3/list_parser.hpp"
 
 #include <chrono>
 #include <cstddef>
