+++ src/include/io/rest/rest_ioctx.hpp
@@ -17,7 +17,7 @@
 #pragma once
 
 #include "io/rest/rest_reactor.hpp"
-#include "io/s3/s3_list_parser.hpp"
+#include "io/rest/s3/list_parser.hpp"
 #include "io/templated_ioctx.hpp"
 
 #include <cstddef>
