+++ src/include/io/rest/s3/sigv4_authorizer.hpp
@@ -16,8 +16,8 @@
 
 #pragma once
 
-#include "io/s3/s3_request_authorizer.hpp"
-#include "io/s3/static_credentials.hpp"
+#include "io/rest/authorizer.hpp"
+#include "io/rest/s3/static_credentials.hpp"
 
 #include <chrono>
 #include <string>
