+++ src/include/io/rest/s3/static_credentials.hpp
