+++ src/io/datasource_factory.cpp
@@ -20,8 +20,8 @@
 #include "io/kvikio/kvikio_context.hpp"
 #include "io/object_store_config.hpp"
 #include "io/rest/rest_ioctx.hpp"
-#include "io/s3/sirius_sigv4_authorizer.hpp"
-#include "io/s3/static_credentials.hpp"
+#include "io/rest/s3/sigv4_authorizer.hpp"
+#include "io/rest/s3/static_credentials.hpp"
 #include "io/uring/uring_ioctx.hpp"
 #include "log/logging.hpp"
 #include "scan_manager/config.hpp"
