+++ src/io/rest/rest_ioctx.cpp
@@ -16,7 +16,7 @@
 
 #include "io/rest/rest_ioctx.hpp"
 
-#include "io/s3/sigv4.hpp"
+#include "io/rest/s3/sigv4.hpp"
 #include "io/uri_parser.hpp"
 
 #include <algorithm>
